+++ docs/app/[lng]/not-found.tsx
@@ -3,7 +3,7 @@
 import dynamic from 'next/dynamic'
 import { usePathname } from 'next/navigation'
 import Link from 'next/link'
-import { Container, Grid, Paper } from '@mui/material'
+import { Container, Grid2 as Grid, Paper } from '@mui/material'
 import { Error as ErrorIcon, Link as LinkIcon } from '@mui/icons-material'
 
 // Components
@@ -21,19 +21,19 @@ const NotFoundPage = () => {
 			<Container maxWidth="sm" sx={{ display: 'flex', alignItems: 'center', height: '100dvh' }}>
 				<Paper variant="outlined" sx={{ width: 1, borderRadius: 2 }}>
 					<Grid container flexDirection="column" alignItems="center" spacing={2} py={4} px={{ xs: 1, md: 2 }}>
-						<Grid item>
+						<Grid>
 							<ErrorIcon color="error" sx={{ fontSize: 100 }} />
 						</Grid>
 
-						<Grid item>
+						<Grid>
 							<TitleTypographyAtom lng={lng} text="common:title.notFound" />
 						</Grid>
 
-						<Grid item>
+						<Grid>
 							<ParagraphTypographyAtom lng={lng} text="common:description.notFound" />
 						</Grid>
 
-						<Grid item>
+						<Grid>
 							<Link href="/">
 								<PrimaryButtonAtom lng={lng} title="form:button.goHome" startIcon={<LinkIcon />} />
 							</Link>